Rename: 신청, 쪽지 모달 PostDetail로 이동
